Exploration of MIT’s Observatory of Economic Complexity
Group Project

Exploration of explore MIT’s Observatory of Economic Complexity (“OEC”), a database which gathers product trade data, e.g. imports and exports of goods, by country for each year.

From this data, the aim was to look at some long-time trends that could be deciphered from ebb and flow of imports and exports in the European region, particularly in the light of major diplomatic and economic integration such as formation of European Monetary Union.